qrGeneration/requirement.txt
pillow==11.1.0
psycopg2==2.9.10
python-dotenv==1.0.1
qrcode==8.0
